requirement.txt
accelerate==0.21.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
msgpack==1.0.2
msgpack-rpc-python==0.4.1
tornado==4.5.3
airsim==1.8.1
altair==5.5.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
bitsandbytes==0.41.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
cmake==3.25.0
contourpy==1.3.1
cycler==0.12.1
decord==0.6.0
deepspeed==0.9.5
defusedxml==0.7.1
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fairscale==0.4.13
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.13.1
flash-attn==2.5.9.post1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gradio==3.35.2
gradio_client==0.2.9
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.27.1
idna==3.4
jinja2==3.0.3
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
latex2mathml==3.77.0
linkify-it-py==2.0.3
lit==15.0.7
llvmlite==0.43.0
markdown-it-py==2.2.0
markdown2==2.5.2
MarkupSafe==2.1.5
matplotlib==3.10.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.21.1
networkx==3.2.1
ninja==1.11.1.3
numba==0.60.0
numpy==1.26.3
openai==1.59.5
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
orjson==3.10.13
packaging==24.2
pandas==2.2.3
peft==0.4.0
pillow==10.2.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pycocotools==2.0.8
pydantic==1.10.20
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.28.1
rpds-py==0.22.3
safetensors==0.5.1
scikit-learn==1.2.2
scipy==1.15.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.19.2
setproctitle==1.3.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.41.3
supervision==0.25.1
svgwrite==1.4.3
sympy==1.13.1
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.13.3
tomli==2.2.1
tqdm==4.67.1
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==1.26.13
uvicorn==0.34.0
wandb==0.18.7
wavedrom==2.0.3.post3
websockets==14.1
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
